+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Version 3.0.20
+- Improved AOE template targeting.
+- Initial AOE EXPLOSION support.  Sorts by range to center of template and shows distance to center.  Damage falloff not implemented yet.  [#151](https://github.com/dmdorman/hero6e-foundryvtt/issues/151)
+- Non PCs are marked as defeated when they drop below -10 STUN.  Once defeated they no longer get post segment 12 recoveries.
+- Improved handling of Mental attacks OMCV/DMCV, DMCV buffs and Mental Combat Skill Levels. [#272](https://github.com/dmdorman/hero6e-foundryvtt/issues/272)
+- Fixed inability to manually create new active effects. [#271](https://github.com/dmdorman/hero6e-foundryvtt/issues/271)
+- Improved attack cards to show all attack modifier tags.
+
 # Version 3.0.19
 - Framework modifiers now transfer REDUCEDEND to slots [#266](https://github.com/dmdorman/hero6e-foundryvtt/issues/266)
 - Improved MULTIPOWER descriptions and slot costs.